#30 Remove deleted member from groups

+++ nyugtafeloszto/src/app/member/list/list.component.ts
@@ -20,6 +20,7 @@ import { Subscription } from 'rxjs';
 
 import { Member } from '../../shared/models/Member';
 import { MemberService } from 'src/app/shared/services/member.service';
+import { GroupService } from 'src/app/shared/services/group.service';
 
 @Component({
   selector: 'app-list',
@@ -44,7 +45,10 @@ export class ListComponent implements OnInit, AfterViewInit, OnDestroy {
   user?: string | null;
   memberSubscription?: Subscription;
 
-  constructor(private memberService: MemberService) {}
+  constructor(
+    private memberService: MemberService,
+    private groupService: GroupService
+  ) {}
 
   ngOnInit(): void {
     this.user = localStorage.getItem('user');
@@ -140,8 +144,18 @@ export class ListComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   deleteMember(member: Member) {
-    if (member.id) {
+    if (member.id && this.user) {
       this.memberService.delete(member.id);
+
+      this.groupService
+        .getByMember(JSON.parse(this.user).uid, member.id)
+        .subscribe((data) => {
+          data.forEach((group) => {
+            group.members = group.members.filter((el) => el !== member.id);
+            this.groupService.update(group);
+          });
+        });
     }
+
   }
 }
